backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
pandas>=2.2.0
pandas-ta>=0.3.14b
numpy>=1.26.0
requests>=2.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
websockets>=13.0
pydantic>=2.9.0
supabase>=2.0.0
